@@ -127,31 +131,32 @@
     "vm-browserify": "0.0.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/runtime": "^7.4.5",
-    "@react-native-community/cli": "2.0.0-rc.2",
+    "@babel/core": "^7.5.5",
+    "@babel/runtime": "^7.5.5",
+    "@react-native-community/cli": "2.8.3",
     "@react-native-community/eslint-config": "^0.0.5",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^8.2.2",
+    "babel-eslint": "^10.0.2",
     "babel-jest": "^24.8.0",
     "babel-plugin-date-fns": "^0.2.1",
     "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-styled-components": "^1.10.6",
     "babel-plugin-rewire": "^1.2.0",
+    "babel-plugin-styled-components": "^1.10.6",
     "babel-plugin-transform-remove-console": "^6.9.4",
     "detox": "^12.8.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-react": "^7.13.0",
+    "eslint": "^6.1.0",
+    "eslint-config-airbnb-base": "^13.2.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-react": "^7.14.3",
     "eslint-plugin-react-native": "^3.7.0",
     "eslint-plugin-react-native-animation-linter": "^0.1.2",
     "jest": "^24.8.0",
-    "metro-react-native-babel-preset": "^0.55.0",
+    "metro-react-native-babel-preset": "^0.56.0",
     "mocha": "^6.1.4",
+    "react-native-bundle-visualizer": "^2.0.1",
     "react-native-clean-project": "^3.1.0",
     "react-test-renderer": "16.8.3",
-    "schedule": "0.4.0"
+    "schedule": "0.5.0"
   },
   "resolutions": {
     "**/eslint-plugin-import": "2.14.0",
