@@ -7,8 +7,8 @@ import FloatingEmoji from './FloatingEmoji';
 
 const getRandomNumber = (min, max) => Math.random() * (max - min) + min;
 
-const createEmojiItem = () => {
-  const right = `${getRandomNumber(0, 80)}%`;
+const createEmojiItem = (range) => {
+  const right = `${getRandomNumber(...range)}%`;
 
   return ({
     id: right,
@@ -21,13 +21,17 @@ export default class FloatingEmojis extends PureComponent {
     color: PropTypes.string,
     count: PropTypes.number,
     distance: PropTypes.number,
+    duration: PropTypes.number,
     emoji: PropTypes.string.isRequired,
+    range: PropTypes.arrayOf(PropTypes.number),
     size: PropTypes.string.isRequired,
     style: stylePropType,
+    top: PropTypes.number,
   }
 
   static defaultProps = {
     count: -1,
+    range: [0, 80],
     size: 'h2',
   }
 
@@ -35,18 +39,20 @@ export default class FloatingEmojis extends PureComponent {
     emojis: [],
   }
 
-  componentWillUpdate(nextProps) {
-    const oldCount = this.props.count;
-    const newCount = nextProps.count;
+  componentDidUpdate(prevProps) {
+    const oldCount = prevProps.count;
+    const newCount = this.props.count;
     const numEmojis = newCount - oldCount;
 
     if (numEmojis <= 0) return;
 
     const items = Array(numEmojis).fill();
-    const newEmojis = items.map((_, i) => oldCount + i).map(createEmojiItem);
+    const newEmojis = items.map((_, i) => oldCount + i).map(this.createItem);
 
     this.setState({ emojis: this.state.emojis.concat(newEmojis) });
   }
+
+  createItem = () => createEmojiItem(this.props.range)
 
   removeEmoji = (id) => {
     const newEmojis = this.state.emojis.filter(emoji => emoji.id !== id);
@@ -63,11 +69,13 @@ export default class FloatingEmojis extends PureComponent {
         <FloatingEmoji
           {...item}
           distance={this.props.distance}
+          duration={this.props.duration}
           emoji={this.props.emoji}
           id={id}
           key={id}
           onComplete={this.removeEmoji}
           size={this.props.size}
+          top={this.props.top}
         />
       ))}
     </View>
