@@ -91,6 +91,7 @@ const hasRowChanged = (r1, r2) => {
 
 class RecyclerAssetList extends Component {
   static propTypes = {
+    externalScrollView: PropTypes.any,
     fetchData: PropTypes.func,
     hideHeader: PropTypes.bool,
     openFamilyTabs: PropTypes.array,
@@ -135,6 +136,7 @@ class RecyclerAssetList extends Component {
       dataProvider: new DataProvider(hasRowChanged, this.getStableId),
       headersIndices: [],
       isRefreshing: false,
+      itemsCount: 0,
     };
 
     this.layoutProvider = new LayoutProvider(
@@ -145,7 +147,7 @@ class RecyclerAssetList extends Component {
           return ViewTypes.HEADER;
         }
 
-        if (index === this.state.length - 1) {
+        if (index === this.state.itemsCount - 1) {
           return ViewTypes.FOOTER;
         }
 
@@ -180,7 +182,7 @@ class RecyclerAssetList extends Component {
                 return {
                   get: ViewTypes.UNIQUE_TOKEN_ROW,
                   isFirst: index === headersIndices[collectiblesIndex] + 1,
-                  isLast: index === this.state.length - 2,
+                  isLast: index === this.state.itemsCount - 2,
                   rowCount: get(sections, `[${collectiblesIndex}].data[${familyIndex}].tokens`, []).length,
                 };
               }
@@ -188,7 +190,7 @@ class RecyclerAssetList extends Component {
             return {
               get: ViewTypes.UNIQUE_TOKEN_ROW_CLOSED,
               isFirst: index === headersIndices[collectiblesIndex] + 1,
-              isLast: index === this.state.length - 2,
+              isLast: index === this.state.itemsCount - 2,
             };
           }
         }
@@ -257,7 +259,7 @@ class RecyclerAssetList extends Component {
       areSmallCollectibles,
       dataProvider: state.dataProvider.cloneWithRows(items),
       headersIndices,
-      length: items.length,
+      itemsCount: items.length,
     };
   }
 
@@ -467,7 +469,13 @@ class RecyclerAssetList extends Component {
   };
 
   render() {
-    const { hideHeader, renderAheadOffset, ...props } = this.props;
+    const {
+      externalScrollView,
+      fetchData,
+      hideHeader,
+      renderAheadOffset,
+      ...props
+    } = this.props;
     const { dataProvider, headersIndices } = this.state;
 
     return (
@@ -476,8 +484,9 @@ class RecyclerAssetList extends Component {
           <RecyclerListView
             {...props}
             dataProvider={dataProvider}
-            extendedState={{ headersIndices }}
+            extendedState={headersIndices}
             itemAnimator={layoutItemAnimator}
+            externalScrollView={externalScrollView}
             layoutProvider={this.layoutProvider}
             onScroll={this.handleScroll}
             ref={this.handleListRef}
@@ -488,7 +497,7 @@ class RecyclerAssetList extends Component {
               top: hideHeader ? 0 : AssetListHeader.height,
             }}
             scrollViewProps={{
-              refreshControl: this.renderRefreshControl(),
+              refreshControl: fetchData && this.renderRefreshControl(),
             }}
             style={{
               backgroundColor: colors.white,
