@@ -1,5 +1,7 @@
 import React from 'react';
 import { Text } from 'react-primitives';
 
-const Nbsp = props => <Text {...props}>&nbsp;</Text>;
+const unicodeValue = '\xa0';
+const Nbsp = props => <Text {...props}>{unicodeValue}</Text>;
+Nbsp.unicode = unicodeValue;
 export default Nbsp;
