@@ -1,52 +1,53 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { PureComponent } from 'react';
 import { TextInput } from 'react-native';
 import { buildTextStyles, colors } from '../../styles';
 
-const Input = ({
-  allowFontScaling,
-  autoCapitalize,
-  autoCorrect,
-  keyboardType,
-  placeholderTextColor,
-  spellCheck,
-  ...props
-}) => (
-  <TextInput
-    {...props}
-    allowFontScaling={allowFontScaling}
-    autoCapitalize={autoCapitalize}
-    autoCorrect={autoCorrect}
-    css={buildTextStyles}
-    keyboardType={keyboardType}
-    placeholderTextColor={placeholderTextColor}
-    spellCheck={spellCheck}
-  />
-);
+export default class Input extends PureComponent {
+  static propTypes = {
+    allowFontScaling: PropTypes.bool,
+    autoCapitalize: PropTypes.string,
+    autoCorrect: PropTypes.bool,
+    keyboardType: PropTypes.string,
+    placeholderTextColor: PropTypes.string,
+    spellCheck: PropTypes.bool,
+  }
 
-Input.propTypes = {
-  allowFontScaling: PropTypes.bool,
-  autoCapitalize: PropTypes.string,
-  autoCorrect: PropTypes.bool,
-  keyboardType: PropTypes.oneOf([
-    'decimal-pad',
-    'default',
-    'email-address',
-    'name-phone-pad',
-    'number-pad',
-    'numeric',
-    'phone-pad',
-  ]),
-  placeholderTextColor: PropTypes.string,
-  spellCheck: PropTypes.bool,
-};
+  static defaultProps = {
+    allowFontScaling: false,
+    autoCapitalize: 'none',
+    autoCorrect: false,
+    placeholderTextColor: colors.placeholder,
+    spellCheck: true,
+  }
 
-Input.defaultProps = {
-  allowFontScaling: false,
-  autoCapitalize: 'none',
-  autoCorrect: false,
-  placeholderTextColor: colors.placeholder,
-  spellCheck: true,
-};
+  ref = React.createRef()
 
-export default Input;
+  focus = event => this.ref.current.focus(event)
+
+  render = () => {
+    const {
+      allowFontScaling,
+      autoCapitalize,
+      autoCorrect,
+      keyboardType,
+      placeholderTextColor,
+      spellCheck,
+      ...props
+    } = this.props;
+
+    return (
+      <TextInput
+        {...props}
+        allowFontScaling={allowFontScaling}
+        autoCapitalize={autoCapitalize}
+        autoCorrect={autoCorrect}
+        css={buildTextStyles}
+        keyboardType={keyboardType}
+        placeholderTextColor={placeholderTextColor}
+        ref={this.ref}
+        spellCheck={spellCheck}
+      />
+    );
+  }
+}
