@@ -1,19 +1,14 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import styled from 'styled-components';
-import { colors, padding } from '../../styles';
+import { colors } from '../../styles';
 import { deviceUtils } from '../../utils';
 import { Centered, Column } from '../layout';
 import TouchableBackdrop from '../TouchableBackdrop';
 
-const Container = styled(Centered)`
-  ${padding(15)};
-  height: 100%;
-`;
-
 const ModalElement = styled(Column)`
   background-color: ${colors.white};
-  border-radius: 12;
+  border-radius: ${({ radius }) => radius || 12};
   flex-shrink: 0;
   height: ${({ height }) => height};
   width: 100%;
@@ -24,27 +19,37 @@ const Modal = ({
   height,
   onCloseModal,
   statusBarStyle,
+  containerPadding,
   ...props
 }) => (
-  <Container direction="column" style={{ justifyContent: fixedToTop ? 'flex-start' : 'center', top: fixedToTop ? 91 : 0 }}>
+  <Centered
+    style={{ justifyContent: fixedToTop ? 'flex-start' : 'center', top: fixedToTop ? 91 : 0 }}
+    direction="column"
+    height="100%"
+    padding={containerPadding}
+    width="100%"
+  >
     <TouchableBackdrop onPress={onCloseModal} />
     <ModalElement
       {...props}
       height={height}
     />
-  </Container>
+  </Centered>
 );
 
 Modal.propTypes = {
   fixedToTop: PropTypes.bool,
+  containerPadding: PropTypes.number.isRequired,
   height: PropTypes.number.isRequired,
-  onCloseModal: PropTypes.func.isRequired,
+  onCloseModal: PropTypes.func,
   statusBarStyle: PropTypes.oneOf(['default', 'light-content', 'dark-content']),
 };
 
 Modal.defaultProps = {
   fixedToTop: false,
+  containerPadding: 15,
   height: deviceUtils.dimensions.height - 230,
+  onCloseModal: () => null,
   statusBarStyle: 'light-content',
 };
 
