@@ -1,3 +1,4 @@
+import { getExecutionDetails } from '@uniswap/sdk';
 import contractMap from 'eth-contract-metadata';
 import { ethers } from 'ethers';
 import { get, map, zipObject } from 'lodash';
@@ -7,11 +8,71 @@ import {
   fromWei,
   multiply,
 } from '../helpers/utilities';
+import { loadWallet } from '../model/wallet';
 import exchangeABI from '../references/uniswap-exchange-abi.json';
 import erc20ABI from '../references/erc20-abi.json';
 import { web3Provider } from './web3';
 
-export default async function getUniswapLiquidityInfo(accountAddress, exchangeContracts) {
+const convertArgsForEthers = (methodArguments) => methodArguments.map(arg => (typeof arg === 'object') ? ethers.utils.bigNumberify(arg.toFixed()) : arg);
+
+const convertValueForEthers = (value) => {
+  const valueBigNumber = ethers.utils.bigNumberify(value.toString());
+  return ethers.utils.hexlify(valueBigNumber);
+};
+
+export const executeSwap = async (tradeDetails) => {
+  const executionDetails = getExecutionDetails(tradeDetails);
+  const wallet = await loadWallet();
+  if (!wallet) return null;
+  const {
+    exchangeAddress,
+    methodArguments,
+    methodName,
+    value: rawValue,
+  } = executionDetails;
+  const exchange = new ethers.Contract(exchangeAddress, exchangeABI, wallet);
+  const updatedMethodArgs = convertArgsForEthers(methodArguments);
+  const value = convertValueForEthers(rawValue);
+  try {
+    switch (methodName) {
+    case 'ethToTokenSwapInput': {
+      const gasLimit = await exchange.estimate.ethToTokenSwapInput(...updatedMethodArgs, { value });
+      return exchange.ethToTokenSwapInput(...updatedMethodArgs, { gasLimit, value });
+    }
+    case 'ethToTokenSwapOutput': {
+      const gasLimit = await exchange.estimate.ethToTokenSwapOutput(...updatedMethodArgs, { value });
+      return exchange.ethToTokenSwapOutput(...updatedMethodArgs, { gasLimit, value });
+    }
+    case 'tokenToEthSwapInput': {
+      // TODO approval check
+      const gasLimit = await exchange.estimate.tokenToEthSwapInput(...updatedMethodArgs, { value });
+      return exchange.tokenToEthSwapInput(...updatedMethodArgs, { gasLimit, value });
+    }
+    case 'tokenToEthSwapOutput': {
+      // TODO approval check
+      const gasLimit = await exchange.estimate.tokenToEthSwapOutput(...updatedMethodArgs, { value });
+      return exchange.tokenToEthSwapOutput(...updatedMethodArgs, { gasLimit, value });
+    }
+    case 'tokenToTokenSwapInput': {
+      // TODO approval check
+      const gasLimit = await exchange.estimate.tokenToTokenSwapInput(...updatedMethodArgs, { value });
+      return exchange.tokenToTokenSwapInput(...updatedMethodArgs, { gasLimit, value });
+    }
+    case 'tokenToTokenSwapOutput': {
+      // TODO approval check
+      const gasLimit = await exchange.estimate.tokenToTokenSwapOutput(...updatedMethodArgs, { value });
+      return exchange.tokenToTokenSwapOutput(...updatedMethodArgs, { gasLimit, value });
+    }
+    default:
+      return null;
+    }
+  } catch (error) {
+    console.log('error exchanging', error);
+    return null;
+  }
+};
+
+export const getUniswapLiquidityInfo = async (accountAddress, exchangeContracts) => {
   const promises = map(exchangeContracts, async (exchangeAddress) => {
     try {
       const ethReserveCall = web3Provider.getBalance(exchangeAddress);
@@ -81,4 +142,4 @@ export default async function getUniswapLiquidityInfo(accountAddress, exchangeCo
 
   const results = await Promise.all(promises);
   return zipObject(exchangeContracts, results);
-}
+};
