@@ -1,33 +1,38 @@
-import Animated from 'react-native-reanimated';
+import produce from 'immer';
+import { Animated } from 'react-native';
 
 const { Value } = Animated;
 
 // -- Constants --------------------------------------- //
-const UPDATE_TRANSITION_PROPS = 'navigation/UPDATE_TRANSITION_PROPS';
+const UPDATE_STACK_TRANSITION_PROPS = 'navigation/UPDATE_STACK_TRANSITION_PROPS';
+const UPDATE_TABS_TRANSITION_PROPS = 'navigation/UPDATE_TABS_TRANSITION_PROPS';
 
-export const updateTransitionProps = (payload) => (dispatch) => {
-  dispatch({ payload, type: UPDATE_TRANSITION_PROPS });
+export const updateStackTransitionProps = (payload) => (dispatch) => {
+  dispatch({ payload, type: UPDATE_STACK_TRANSITION_PROPS });
+};
+
+export const updateTabsTransitionProps = (payload) => (dispatch) => {
+  dispatch({ payload, type: UPDATE_TABS_TRANSITION_PROPS });
 };
 
 // -- Reducer ----------------------------------------- //
 const INITIAL_STATE = {
-  transitionProps: {
+  stackTransitionProps: {
     isTransitioning: false,
     position: new Value(0),
   },
+  tabsTransitionProps: {
+    isTransitioning: false,
+  },
 };
 
-export default (state = INITIAL_STATE, action) => {
-  switch (action.type) {
-  case UPDATE_TRANSITION_PROPS:
-    return {
-      ...state,
-      transitionProps: {
-        ...state.transitionProps,
-        ...action.payload,
-      },
-    };
-  default:
-    return state;
-  }
-};
+export default (state = INITIAL_STATE, action) => (
+  produce(state, draft => {
+    if (action.type === UPDATE_STACK_TRANSITION_PROPS) {
+      Object.assign(draft.stackTransitionProps, action.payload);
+    }
+    if (action.type === UPDATE_TABS_TRANSITION_PROPS) {
+      Object.assign(draft.tabsTransitionProps, action.payload);
+    }
+  })
+);
