@@ -13,17 +13,10 @@ import {
 import { AssetList } from '../components/asset-list';
 import BlurOverlay from '../components/BlurOverlay';
 import { FabWrapper } from '../components/fab';
-import {
-  CameraHeaderButton,
-  Header,
-  ProfileHeaderButton,
-} from '../components/header';
+import { CameraHeaderButton, Header, ProfileHeaderButton } from '../components/header';
 import { Page } from '../components/layout';
+import { getShowShitcoinsSetting, updateShowShitcoinsSetting } from '../handlers/commonStorage';
 import buildWalletSectionsSelector from '../helpers/buildWalletSections';
-import {
-  getShowShitcoinsSetting,
-  updateShowShitcoinsSetting,
-} from '../handlers/commonStorage';
 import {
   withAccountData,
   withAccountSettings,
@@ -31,12 +24,12 @@ import {
   withDataInit,
   withIsWalletEmpty,
   withIsWalletEthZero,
-  withUniqueTokens,
   withStatusBarStyle,
+  withUniqueTokens,
   withUniswapLiquidity,
 } from '../hoc';
 import { position } from '../styles';
-import { isNewValueForPath } from '../utils';
+import { deviceUtils, isNewValueForPath } from '../utils';
 
 class WalletScreen extends Component {
   static propTypes = {
@@ -91,8 +84,7 @@ class WalletScreen extends Component {
       || isNewCurrency
       || isNewBlurIntensity
       || isNewSections
-      || isNewShowShitcoins
-      || isNewTransitionProps;
+      || isNewShowShitcoins;
   }
 
   render = () => {
