@@ -24,7 +24,7 @@ export default css`
       : ''
   )}
   color: ${({ color }) => (colors.get(color) || colors.dark)}
-  font-family: ${({ family }) => fonts.family[family || 'SFProText']};
+  font-family: ${({ family = 'SFProText', mono }) => fonts.family[mono ? 'SFMono' : family]};
   font-size: ${({ size }) => fonts.size[size || 'medium']};
   font-weight: ${({ weight }) => fonts.weight[weight || 'regular']};
 `;
