@@ -7,12 +7,15 @@ const base = {
   blueGreyDark: '#3C4252', // '60, 66, 82'
   blueGreyDarker: '#0F0F11', // '15, 15, 17'
   blueGreyLight: '#A1A5AC',
-  blueGreyLightest: '#8A8E97', // '138, 142, 151'
-  blueGreyLigter: '#666A73', // '102, 106, 115'
+  blueGreyLighter: '#666A73', // '102, 106, 115'
+  blueGreyLightest: '#888D96', // '136, 141, 150'
   blueGreyMedium: '#636875', // '99, 104, 117'
   blueGreyMediumLight: '#7b7f8a', // '123, 127, 138'
+  brightOrange: '#FFB624', // '255, 182, 36'
+  brightRed: '#FF4B19', // '255, 75, 25'
   dark: '#25292E', // '37, 41, 46'
   darkGrey: '#71778a', // '113, 119, 138'
+  dodgerBlue: '#575CFF', // '87, 92, 255'
   green: '#00994d', // '0, 153, 77'
   grey: '#a9adb9', // '169, 173, 185'
   highlightBackground: '#F0F7FF', // '240, 247, 255'
